Monitoring-real-time-Road-Traffic/requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
happybase==1.2.0
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kafka==1.3.5
MarkupSafe==2.1.5
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
ply==3.11
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
thriftpy2==0.5.1
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
